20.02.25) step1 리팩토링
- 불필요한 IndexController 제거
- LocalDateTime 사용
- id의 타입을 int에서 Long으로 변경
- 회원 정보 수정 시 post를 put으로 변경 실패.

+++ src/main/java/com/codessquad/qna/QuestionController.java
@@ -11,6 +11,7 @@ import java.util.List;
 
 @Controller
 public class QuestionController {
+
     private List<Question> questions = new ArrayList<>();
 
     @GetMapping("/qna/form")
@@ -20,7 +21,7 @@ public class QuestionController {
 
     @PostMapping("/questions")
     public String questions(Question question) {
-        question.setId(questions.size()+1);
+        question.setId((long) (questions.size() + 1));
         questions.add(question);
         System.out.println("question : "+ question);
         return "redirect:/";
